bug fixes
Fixed some bugs surrounding promotions and viewing game history.
Fixed a bug where a promoted piece checkmating the enemy would be demoted when viewing in the turn history panel due to the turn history panel overriding the turn to be an invalid one on game end. Prevented the history from setting a state that has already been set.
Fixed a bug where the turn number of a promotion for the black team was being recorded incorrectly. 
Fixed a bug where a promote piece won't visually update when loading a game that ended while it was promoted and still on the board.